clean up and comments

+++ connectfour.c
@@ -29,6 +29,8 @@ const unsigned char sage22[] ={
  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xC3, 0x0E, 0x00, 0x00, 0xC3, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x3C, 0x37, 0x52, 0x9B, 0x3A, 0x36, 0x54, 0xAA, 0x3B, 0x67,
  0x92, 0xAC, 0x3F, 0x6E, 0x99, 0xCE, 0xFF,
 };
+
+
 	
 #define GPIO_PORTF_DATA_R       (*((volatile unsigned long *)0x400253FC))
 #define GPIO_PORTF_DIR_R        (*((volatile unsigned long *)0x40025400))
@@ -43,9 +45,9 @@ const unsigned char sage22[] ={
 
 
 /********consts*******/	
-const int BOARDWIDTH = 13;
-const int BOARDHIGHT = 8;
-const int CELLSPACING = 6;
+const short BOARDWIDTH = 13;
+const short BOARDHIGHT = 8;
+const short CELLSPACING = 6;
 
 /*******endgamemsgs*****/
 char sqwinmsg[] = "SQUARE WINS";
@@ -54,36 +56,36 @@ char tiemsg[]      = "TIE!";
 char frmsg[]    = "you still friends";
 
 /*******singeltons*********/
-int sage     = 0;     // singelton determines which column will be played in.
-int turn     = 1;     // 0->first player , 1 -> secondplayer
-int gameover = 0;     //0 -> still in play. 1->game over.
-int board[BOARDWIDTH][BOARDHIGHT] = {0};  // 2d-array represent the playfield, 0 -> empty
-int winner   = 0;
+short sage     = 0;     // determines which column will be played in.
+short turn     = 1;     // 0->first player , 1 -> secondplayer
+short gameover = 0;     //0 -> still in play. 1->game over.
+short board[BOARDWIDTH][BOARDHIGHT] = {0};  // 2d-array represent the playfield, 0 -> empty
+short winner   = 0;
 
 /*****debug singeltons*****/
-int debugvalid;
-int debugplayed = 0; 
+short debugvalid;
+short debugplayed = 0; 
 
 /*******switches hacking******/
 unsigned long SW1,SW2;
 void PortF_Init(void);
-int read_sw1() { return GPIO_PORTF_DATA_R&0x10;}
-int read_sw2() { return GPIO_PORTF_DATA_R&0x01;}
-int pressedsw2 = 0;
+short read_sw1() { return GPIO_PORTF_DATA_R&0x10;}
+short read_sw2() { return GPIO_PORTF_DATA_R&0x01;}
+short pressedsw2 = 0;
 
 /****************** declared fucntions; gotta love c ********/
 void move_sage(void);
 void display_board(void);
 void display_sage(void);
-void display_chipdrop(int,int); 
+void display_chipdrop(short,short); 
 void move_sage(void);
-int put(int);
-int toggleturn(int);
-int play(void);
-int getcell(int,int);
-int check_cell(int,int,int);
-int check_board(int);
-int valid(int,int);
+short put(short);
+short toggleturn(short);
+short play(void);
+short getcell(short,short);
+short check_cell(short,short,short);
+short check_board(short);
+short valid(short,short);
 
 int main(void){	
     TExaS_Init(SSI0_Real_Nokia5110_Scope);
@@ -93,7 +95,6 @@ int main(void){
     while(!gameover && !winner){
         if(!read_sw1())
 	        move_sage();
-        
         if(!pressedsw2 && !read_sw2()&& read_sw1()){
                 winner = play();      //winner is 0 untill someone wins
                 pressedsw2=1;
@@ -155,25 +156,27 @@ void move_sage(void){
 	sage = (sage+1)%BOARDWIDTH;
 }
 
-int put(int turn){
+short put(short turn){
 	// try to insert a chip at the top of col pointed to by
 	// sage, if the col is full return -1, otherwise return the
 	// row where the chip was inserted.
-	int row = 0;
+	short row = 0;
     while( !board[sage][row] && row < BOARDHIGHT){
 	    display_chipdrop(row,turn);	
         ++row;
     }
-	if(!row)  //this column is full, cannot insert chip.
+    if(!row)  //this column is full, cannot insert chip.
 		return -1;
 	board[sage][row-1] = turn; // set the cell to 1 or 2
 	return row;
 }
 
-int play(void){
-    // what everytime someone try to make a move (ie: play)
-    // triggered everytiem  sw2 is pressed.
-    int valid = put(turn+1);
+short play(void){
+    // triggerd everytime someone try to make a move (ie: play)
+    // triggered everytime  sw2 is pressed.
+    // return 0 if there is no winner after this move
+    // other wise returns the winner number (1 or 2).
+    short valid = put(turn+1); 
     debugvalid = valid;
     if(valid ==-1)
         return 0;
@@ -184,14 +187,15 @@ int play(void){
     return 0;
 }
 
-int toggleturn(int turn){return (turn+1)%2;} // switch between players 0 and 1
+short toggleturn(short turn){return (turn+1)%2;} // switch between players 0 and 1
+
 /*********************** Display funcs ****************/ 
 void display_board(){
 	// display the board on the srcreen.
-	int col,row;
+	short col,row;
 	for(col=0;col<BOARDHIGHT;col++){
         for(row=0;row<BOARDWIDTH;row++){
-		    int x = board[row][col];
+		    short x = board[row][col];
 		    if(x==1)
 		    	Nokia5110_PrintBMP((row+1) * CELLSPACING, 5*(col+2), cursor , 0); //draw a cross
 		    if(x==2)
@@ -201,12 +205,15 @@ void display_board(){
 }
 
 void display_sage(void){
-    //display sage in the correct position according to sageposx
-    int posy = 5;
-    int posx = (sage+1) * CELLSPACING;
+    //display sage in the correct position according to the variabl sage
+    short posy = 5;
+    short posx = (sage+1) * CELLSPACING;
 	Nokia5110_PrintBMP(posx,posy, cursor , (turn?14:0));
 }
-void display_chipdrop(int row,int turn){
+void display_chipdrop(short row,short turn){
+    // animation that shows a chip being played into 
+    // place, by moving one row at a time
+    // it so simple, but i really like it
     Nokia5110_ClearBuffer();
     board[sage][row] = turn;
     display_board();
@@ -217,16 +224,16 @@ void display_chipdrop(int row,int turn){
 }
 
 /******** check for win funcs********/
-int valid(int A,int B){
+short valid(short A,short B){
     //return 1 if (A,B) is a valid cell(ie: inside the board)
     //0 otherwise.
     return A < BOARDWIDTH && B < BOARDHIGHT;
 }
 
-int check_board(int trgt){
+short check_board(short trgt){
     //return 1 if the last move is a winning move, 0 otherwise.
     //triggerd every time a valid move is played.
-    int col = 0, row = 0;
+    short col = 0, row = 0;
 	for(col=0;col<BOARDHIGHT;col++){
         for(row=0;row<BOARDWIDTH;row++){
             if(board[row][col] == trgt)
@@ -237,29 +244,35 @@ int check_board(int trgt){
     return 0;
 }
 
-int check_cell(int row,int col,int trgt){
+short check_cell(short row,short col,short trgt){
     //return 1 if the current cell is the start of
     //a winning sequence
-    int okrow = 1;
-    int okcol = 1;
-    int okbil = 1;
-    int okbir = 1;
-    int i;
-    for(i=1;i<4;i++){
-        if(getcell(row+i,col  )!=trgt)
+
+    // by the end if only of these values is still
+    // 1 then this cell is winning
+    // each arrow represent the direction that function checks
+    // i forgot the word daigonal, so i went with bil XD
+
+    short okrow = 1;
+    short okcol = 1;
+    short okbil = 1; 
+    short okbir = 1;
+    short i;
+    for(i=1;i<4;i++){ 
+        if(getcell(row+i,col  )!=trgt) // >
             okrow = 0;
-        if(getcell(row  ,col+i)!=trgt)
+        if(getcell(row  ,col+i)!=trgt) // ^
             okcol = 0;
-        if(getcell(row+i,col+i)!=trgt)
+        if(getcell(row+i,col+i)!=trgt) // ^>
             okbil = 0;
-        if(getcell(row+i,col-i)!=trgt)
+        if(getcell(row+i,col-i)!=trgt) // <^
             okbir = 0;
     }
     if(okrow+okcol+okbil+okbir) return 1;
         return 0;
 }
 
-int getcell(int row,int col){
+short getcell(short row,short col){
     //return the content of a cell, if it is 
     //not on the board return 0
     if(valid(row,col)) return board[row][col];
